Fixed with phpcs + phpstan Added last documentation

+++ public/products/productDelete.php
@@ -1,4 +1,9 @@
 <?php
+
+/**
+ * Script to delete product from table
+ */
+
 require __DIR__ . '/../../vendor/autoload.php';
 
 use Configure\DotEnv;
@@ -8,5 +13,7 @@ use Entity\ProductsMapper;
 
 $productMapper = new ProductsMapper();
 $product = $productMapper->findByKey((int)$_REQUEST['id']);
-$product = $productMapper->delete($product);
+if (!is_null($product)) {
+    $product = $productMapper->delete($product);
+}
 header('Location: ../index.php');
